TIP-452: Fix mob reviews
