+++ src/Pim/Bundle/CatalogBundle/Command/QueryProductCommand.php
@@ -5,6 +5,7 @@ namespace Pim\Bundle\CatalogBundle\Command;
 use Akeneo\Component\StorageUtils\Cursor\CursorInterface;
 use Pim\Component\Catalog\Query\ProductQueryBuilderInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Helper\HelperInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -74,6 +75,8 @@ class QueryProductCommand extends ContainerAwareCommand
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $filters = json_decode($input->getArgument('json_filters'), true);
+        $this->warnDeprecatedMethod($filters);
+
         $pageSize = $input->getOption('page-size');
         $productQueryBuilder = $this->getProductQueryBuilder($filters);
         $products = $productQueryBuilder->execute();
@@ -113,7 +116,7 @@ class QueryProductCommand extends ContainerAwareCommand
      * @param CursorInterface $products
      * @param int             $maxRows
      *
-     * @return \Symfony\Component\Console\Helper\HelperInterface
+     * @return HelperInterface
      */
     protected function buildTable(CursorInterface $products, $maxRows)
     {
@@ -142,8 +145,27 @@ class QueryProductCommand extends ContainerAwareCommand
      */
     protected function getProductQueryBuilder(array $filters)
     {
-        $factory = $this->getContainer()->get('pim_catalog.query.filtered_product_query_builder_factory');
+        $factory = $this->getContainer()->get('pim_catalog.query.product_query_builder_factory');
 
         return $factory->create(['filters' => $filters]);
+    }
+
+    /**
+     * This temporary method warn the user for using deprecated argument format.
+     *
+     * @deprecated Will be removed in 1.7
+     *
+     * @param array $filters
+     */
+    protected function warnDeprecatedMethod(array $filters)
+    {
+        foreach ($filters as $filter) {
+            if (array_key_exists('locale', $filter) || array_key_exists('scope', $filter)) {
+                throw new \InvalidArgumentException(
+                    'The used filter format is deprecated. From version 1.6, the options "scope" and "locale" '.
+                    'must be declared in the "context" option.'
+                );
+            }
+        }
     }
 }
