+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/cleaners.yml
@@ -1,6 +1,5 @@
 parameters:
     pim_enrich.connector.item.mass_edit.temporary_file_cleaner.class: Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\TemporaryFileCleaner
-    pim_enrich.connector.item.mass_edit.variant_group_cleaner.class:  Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\VariantGroupCleaner
 
 services:
     pim_enrich.connector.item.mass_edit.temporary_file_cleaner:
@@ -10,13 +9,3 @@ services:
             - '@pim_catalog.validator.product'
             - '@pim_catalog.repository.product_mass_action'
             - '@pim_catalog.repository.attribute'
-
-    pim_enrich.connector.item.mass_edit.variant_group_cleaner:
-        class: '%pim_enrich.connector.item.mass_edit.variant_group_cleaner.class%'
-        arguments:
-            - '@pim_catalog.query.filtered_product_query_builder_factory'
-            - '@pim_enrich.product.paginator'
-            - '@akeneo_storage_utils.doctrine.object_detacher'
-            - '@pim_catalog.repository.group'
-            - '@pim_catalog.repository.product'
-            - '@translator'
