+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/readers.yml
@@ -6,12 +6,16 @@ services:
     pim_enrich.connector.reader.mass_edit.variant_group_product:
         class: '%pim_enrich.connector.reader.mass_edit.variant_group_product.class%'
         arguments:
-            - '@pim_catalog.query.filtered_product_query_builder_factory'
+            - '@pim_catalog.query.product_query_builder_factory'
             - '@pim_catalog.repository.channel'
             - '@pim_catalog.manager.completeness'
             - '@pim_catalog.converter.metric'
             - true
-            - '@pim_enrich.connector.item.mass_edit.variant_group_cleaner'
+            - '@pim_enrich.product.paginator'
+            - '@akeneo_storage_utils.doctrine.object_detacher'
+            - '@pim_catalog.repository.group'
+            - '@pim_catalog.repository.product'
+            - '@translator'
 
     pim_enrich.connector.reader.mass_edit.family:
         class: '%pim_enrich.connector.reader.mass_edit.family.class%'
